+++ tests/core/test_finetuning.py
@@ -18,7 +18,7 @@ from torch import nn
 from torch.nn import functional as F
 
 from flash import ClassificationTask, Trainer
-from flash.core.finetuning import NoFreeze
+from flash.core.finetuning import FlashBaseFinetuning
 from tests.core.test_model import DummyDataset
 
 
@@ -32,12 +32,9 @@ def test_finetuning(tmpdir: str, strategy):
     task = ClassificationTask(model, F.nll_loss)
     trainer = Trainer(fast_dev_run=True, default_root_dir=tmpdir)
     if strategy == "cls":
-        strategy = NoFreeze()
+        strategy = FlashBaseFinetuning()
     if strategy == 'chocolat':
         with pytest.raises(MisconfigurationException, match="strategy should be within"):
-            trainer.finetune(task, train_dl, val_dl, strategy=strategy)
-    elif strategy is None:
-        with pytest.raises(MisconfigurationException, match="strategy should"):
             trainer.finetune(task, train_dl, val_dl, strategy=strategy)
     else:
         trainer.finetune(task, train_dl, val_dl, strategy=strategy)